construct_geom

From a cppRosm data table, make the geometries.

Usage

construct_geom(cpprnet_dt, complete = FALSE, cores = 1)

Arguments

cpprnet_dt

a data table obtained with cppRosm extract_data()

complete

bool. TRUE to reconstruct Polygons, else use just points.

cores

integer. Run the computation on multiple cores using parallel.

Value

Modifies the provided cpprnet_dt object and adds an sf geometry column to it, with either full geometries or just the centroids. Full geometries typically contain polygons and points.

Examples

cpprnet_filepath <- system.file(package = 'cppRosm','extdata','monaco-latest.osm.pbf')
cpprnet_dt <- cppRosm::extract_data(filename = cpprnet_filepath)
#> → Using main keys `amenity`,`shop`,`tourism`.

cppRosm::construct_geom(cpprnet_dt,complete=TRUE)
#>               id     key   value      lon      lat      attrs
#>           <char>  <char>  <char>    <num>    <num>     <list>
#>    1:   25191432 amenity parking 7.427108 43.73128  <list[7]>
#>    2:   25230434 amenity    fuel 7.412094 43.72851  <list[8]>
#>    3:   25239190 amenity parking 7.429193 43.74531  <list[2]>
#>    4:   25239191 amenity parking 7.417921 43.73090  <list[6]>
#>    5:  267885777 tourism   hotel 7.427918 43.74094 <list[14]>
#>   ---                                                        
#> 1270: 1203849340 amenity parking 7.419638 43.73457  <list[1]>
#> 1271: 1203861575 amenity parking 7.430081 43.74022  <list[1]>
#> 1272: 1229465606 amenity shelter 7.425376 43.73149  <list[3]>
#> 1273: 1229465607 amenity shelter 7.425231 43.73145  <list[3]>
#> 1274: 1229465608 amenity shelter 7.417707 43.73074  <list[3]>
#>                             geometry
#>                       <sfc_GEOMETRY>
#>    1:      POINT (7.427108 43.73128)
#>    2:      POINT (7.412094 43.72851)
#>    3:      POINT (7.429193 43.74531)
#>    4:       POINT (7.417921 43.7309)
#>    5:      POINT (7.427918 43.74094)
#>   ---                               
#> 1270: POLYGON ((7.419592 43.73468...
#> 1271: POLYGON ((7.430036 43.74016...
#> 1272: POLYGON ((7.425359 43.73148...
#> 1273: POLYGON ((7.425274 43.73146...
#> 1274: POLYGON ((7.417716 43.73075...

head(cpprnet_dt$geometry)
#> NULL