Extracting data from raw osm.

Getting non-network data from raw osm files.

Usage

extract_data(filename, main_keys = NULL)

Arguments

filename

a local osm extract file with .osm or .osm.pbf extension.

main_keys

a vector containing main keys to extract the values for. Check osm map features or vignette for a full detail.

Value

a data table with several main columns, including one called attrs which contains lists with named key=value pairs of the secondary osm tags.

Examples

local_osm_filename <- system.file("extdata",'monaco-latest.osm.pbf', package = "cppRosm")

data <- cppRosm::extract_data(local_osm_filename)
#> → Using main keys `amenity`,`shop`,`tourism`.
head(data)
#>           id     key   value      lon      lat      attrs
#>       <char>  <char>  <char>    <num>    <num>     <list>
#> 1:  25191432 amenity parking 7.427108 43.73128  <list[7]>
#> 2:  25230434 amenity    fuel 7.412094 43.72851  <list[8]>
#> 3:  25239190 amenity parking 7.429193 43.74531  <list[2]>
#> 4:  25239191 amenity parking 7.417921 43.73090  <list[6]>
#> 5: 267885777 tourism   hotel 7.427918 43.74094 <list[14]>
#> 6: 267885802 amenity    bank 7.428223 43.74155  <list[9]>
head(data$attrs)
#> [[1]]
#> [[1]]$access
#> [1] "yes"
#> 
#> [[1]]$`addr:country`
#> [1] "MC"
#> 
#> [[1]]$fee
#> [1] "yes"
#> 
#> [[1]]$name
#> [1] "Parking du Chemin des Pêcheurs"
#> 
#> [[1]]$parking
#> [1] "multi-storey"
#> 
#> [[1]]$`parking:condition:area`
#> [1] "ticket"
#> 
#> [[1]]$website
#> [1] "https://documents.monaco-parkings.mc//park/_upl_12_Pecheurs.jpg"
#> 
#> 
#> [[2]]
#> [[2]]$`addr:country`
#> [1] "MC"
#> 
#> [[2]]$brand
#> [1] "Esso"
#> 
#> [[2]]$`brand:wikidata`
#> [1] "Q867662"
#> 
#> [[2]]$`fuel:diesel`
#> [1] "yes"
#> 
#> [[2]]$`fuel:e85`
#> [1] "yes"
#> 
#> [[2]]$`fuel:octane_95`
#> [1] "yes"
#> 
#> [[2]]$`fuel:octane_98`
#> [1] "yes"
#> 
#> [[2]]$name
#> [1] "Esso"
#> 
#> 
#> [[3]]
#> [[3]]$`addr:country`
#> [1] "MC"
#> 
#> [[3]]$parking
#> [1] "surface"
#> 
#> 
#> [[4]]
#> [[4]]$access
#> [1] "yes"
#> 
#> [[4]]$`addr:country`
#> [1] "MC"
#> 
#> [[4]]$fee
#> [1] "yes"
#> 
#> [[4]]$layer
#> [1] "-1"
#> 
#> [[4]]$name
#> [1] "Parking du centre commercial"
#> 
#> [[4]]$parking
#> [1] "underground"
#> 
#> 
#> [[5]]
#> [[5]]$`addr:city`
#> [1] "Monaco"
#> 
#> [[5]]$`addr:country`
#> [1] "MC"
#> 
#> [[5]]$`addr:housenumber`
#> [1] "4"
#> 
#> [[5]]$`addr:postcode`
#> [1] "98000"
#> 
#> [[5]]$`addr:street`
#> [1] "Avenue de la Madone"
#> 
#> [[5]]$email
#> [1] "metropole@metropole.com"
#> 
#> [[5]]$facebook
#> [1] "https://www.facebook.com/HotelMetropoleMonteCarlo/"
#> 
#> [[5]]$internet_access
#> [1] "wlan"
#> 
#> [[5]]$`internet_access:fee`
#> [1] "no"
#> 
#> [[5]]$name
#> [1] "Metropole"
#> 
#> [[5]]$phone
#> [1] "+377 93 15 15 15"
#> 
#> [[5]]$stars
#> [1] "5"
#> 
#> [[5]]$website
#> [1] "https://www.metropole.com/"
#> 
#> [[5]]$wikidata
#> [1] "Q30681475"
#> 
#> 
#> [[6]]
#> [[6]]$`addr:city`
#> [1] "Monte-Carlo"
#> 
#> [[6]]$`addr:country`
#> [1] "MC"
#> 
#> [[6]]$`addr:postcode`
#> [1] "98000"
#> 
#> [[6]]$`addr:street`
#> [1] "Avenue de Grande-Bretagne"
#> 
#> [[6]]$brand
#> [1] "UBS"
#> 
#> [[6]]$`brand:wikidata`
#> [1] "Q193199"
#> 
#> [[6]]$`brand:wikipedia`
#> [1] "en:UBS"
#> 
#> [[6]]$construction
#> [1] "yes"
#> 
#> [[6]]$name
#> [1] "UBS"
#> 
#>